♻️ Refactor Mongo utils
