+++ modules/fantasy-battle/src/Commands/createCharacter.test.ts
@@ -3,7 +3,7 @@ import PcModel from "Models/PlayerCharacter"
 import PlayerUserModel from "Models/PlayerUser"
 import { mockDmMessage, mockPlayerMessage } from "Utils/mockMessage"
 
-import { useDbConnection } from "Utils/mongoTest"
+import { useDbConnection } from "Utils/Mongo/mongoTest"
 import { test, execute } from "./createCharacter"
 
 describe("Command: createCharacter", () => {
