+++ modules/fantasy-battle/src/Commands/listChars.test.ts
@@ -4,7 +4,7 @@ import PcModel from "Models/PlayerCharacter"
 import PlayerUserModel from "Models/PlayerUser"
 
 import { mockPlayerMessage } from "Utils/mockMessage"
-import { useDbConnection } from "Utils/mongoTest"
+import { useDbConnection } from "Utils/Mongo/mongoTest"
 
 import { test, execute } from "./listChars"
