+++ modules/fantasy-battle/src/Models/ExampleScroll/methods/getNext.test.ts
@@ -1,4 +1,4 @@
-import { useDbConnection } from "Utils/mongoTest"
+import { useDbConnection } from "Utils/Mongo/mongoTest"
 import { Scroll } from "../types"
 import ScrollModel from ".."
 import { getNextMethod } from "./getNext"
