+++ modules/fantasy-battle/src/Models/PcResource/index.test.ts
@@ -1,6 +1,6 @@
 import ResourceModel, { Resource } from "./index"
 
-import { useDbConnection } from "Utils/mongoTest"
+import { useDbConnection } from "Utils/Mongo/mongoTest"
 import PcModel from "Models/PlayerCharacter"
 
 describe("PcResource", () => {
