+++ modules/fantasy-battle/src/Models/PlayerCharacter/index.test.ts
@@ -1,6 +1,6 @@
 import ResourceModel from "Models/PcResource"
 import mockAttributes from "Utils/mockAttributes"
-import { useDbConnection } from "Utils/mongoTest"
+import { useDbConnection } from "Utils/Mongo/mongoTest"
 import PcModel, { Pc } from "./index"
 import { createPcProps } from "./statics/create"
