+++ modules/fantasy-battle/src/Models/PlayerCharacter/statics/create.test.ts
@@ -1,4 +1,4 @@
-import { useDbConnection } from "Utils/mongoTest"
+import { useDbConnection } from "Utils/Mongo/mongoTest"
 
 import create, { createPcProps } from "./create"
 import PcModel from ".."
