+++ modules/fantasy-battle/src/Models/PlayerUser/statics/getOrCreate.test.ts
@@ -1,4 +1,4 @@
-import { useDbConnection } from "Utils/mongoTest"
+import { useDbConnection } from "Utils/Mongo/mongoTest"
 import PcModel from "Models/PlayerCharacter"
 import PlayerUserModel from ".."
